backend-app/requirements/base.txt
Django==2.1.2
djangorestframework==3.9.0
gunicorn==19.9.0
mysqlclient==1.3.13
